11/29 기능 추가 및 수정 1. MapManager에서 랜덤길이 선택될때 확률을 조정함 2. 카메라에 state machine 적용함 Idle, Move, Zoom, Rotate 3. 타워 설치카드를 아래에 배치함 4. 화면 위아래로 UI가 나오고 이는 위아래로 숨길 수 있다. 5. 글자가 필요한 UI가 활성화될 때마다 Localization체크하던것에 프레임드랍이 있어 Dictionary로 언어가 바뀔때 한번만 바꾸고 Dictionary에서 가져오는 식으로 프레임 드랍을 없앰.

+++ TowerDefense/Assets/Scripts/DataControl/EnemyData.cs
@@ -9,14 +9,14 @@ namespace DataControl
         public EnemyPoolObjectKey EnemyKey => enemyKey;
         public float Speed => speed;
         public ushort EnemyCoin => enemyCoin;
-        public byte AttackDelay => attackDelay;
+        public float AttackDelay => attackDelay;
         public ushort Damage => damage;
         public int Health => health;
 
         [SerializeField] private EnemyPoolObjectKey enemyKey;
         [Range(0, 5)] [SerializeField] private float speed;
         [SerializeField] private ushort enemyCoin;
-        [SerializeField] private byte attackDelay;
+        [SerializeField] private float attackDelay;
         [SerializeField] private ushort damage;
         [SerializeField] private int health;
     }
